abstract compiler worker generation into a common singleton worker
